requirements.txt
groq
google-generativeai

python-dotenv

pyaudio
SpeechRecognition
numpy

langchain-google-genai
browser-use